# storage/api-docs/api-docs.json
{
    "openapi": "3.0.0",
    "info": {
        "title": "HealthDesk Plus API",
        "description": "Comprehensive healthcare management system for doctors, patients, appointments, prescriptions, and medical records.",
        "contact": {
            "name": "HealthDesk Plus Support",
            "email": "support@healthdeskplus.com"
        },
        "license": {
            "name": "MIT",
            "url": "https://opensource.org/licenses/MIT"
        },
        "version": "1.0.0"
    },
    "servers": [
        {
            "url": "http://localhost:8000/api",
            "description": "Local Development Server"
        },
        {
            "url": "https://api.healthdeskplus.com/api",
            "description": "Production Server"
        }
    ],
    "paths": {
...
        "/ai/onboarding/chat": {
            "post": {
                "tags": [
                    "AI Onboarding Assistant"
                ],
                "summary": "Chat with AI onboarding assistant",
                "description": "Have a conversational interaction with the AI assistant to complete onboarding",
                "operationId": "8343f54c674520decc65668640af055e",
                "requestBody": {
                    "required": true,
                    "content": {
                        "application/json": {
                            "schema": {
                                "required": [
                                    "message"
                                ],
                                "properties": {
                                    "message": {
                                        "description": "User's message to the AI assistant",
                                        "type": "string",
                                        "example": "Hi, I'm Dr. John Doe"
                                    },
                                    "context": {
                                        "description": "Additional context for the conversation",
                                        "type": "object"
                                    }
                                },
                                "type": "object"
                            }
                        }
                    }
                },
                "responses": {
                    "200": {
                        "description": "AI response generated successfully",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "properties": {
                                        "success": {
                                            "type": "boolean",
                                            "example": true
                                        },
                                        "data": {
                                            "properties": {
                                                "response": {
                                                    "type": "string",
                                                    "example": "Hello Dr. Doe! Nice to meet you. I'm here to help you complete your registration."
                                                },
                                                "form_data": {
                                                    "description": "Extracted form data from the conversation",
                                                    "type": "object"
                                                },
                                                "next_action": {
                                                    "properties": {
                                                        "type": {
                                                            "type": "string",
                                                            "example": "update_profile"
                                                        },
                                                        "step": {
                                                            "type": "integer",
                                                            "example": 1
                                                        },
                                                        "message": {
                                                            "type": "string",
                                                            "example": "Updating profile with extracted information"
                                                        }
                                                    },
                                                    "type": "object"
                                                },
                                                "current_step": {
                                                    "type": "integer",
                                                    "example": 1
                                                },
                                                "completion_percentage": {
                                                    "type": "integer",
                                                    "example": 25
                                                },
                                                "suggestions": {
                                                    "type": "array",
                                                    "items": {
                                                        "type": "string"
                                                    },
                                                    "example": [
                                                        "Tell me about your medical qualification",
                                                        "What is your clinic name?"
                                                    ]
                                                },
                                                "requires_confirmation": {
                                                    "type": "boolean",
                                                    "example": false
                                                }
                                            },
                                            "type": "object"
                                        }
                                    },
                                    "type": "object"
                                }
                            }
                        }
                    },
                    "422": {
                        "description": "Validation error",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "properties": {
                                        "success": {
                                            "type": "boolean",
                                            "example": false
                                        },
                                        "message": {
                                            "type": "string",
                                            "example": "Validation errors"
                                        },
                                        "errors": {
                                            "type": "object"
                                        }
                                    },
                                    "type": "object"
                                }
                            }
                        }
                    },
                    "401": {
                        "description": "Unauthorized",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "$ref": "#/components/schemas/UnauthorizedError"
                                }
                            }
                        }
                    }
                },
                "security": [
                    {
                        "bearerAuth": []
                    }
                ]
            }
        },
        "/ai/onboarding/suggestions": {
            "get": {
                "tags": [
                    "AI Onboarding Assistant"
                ],
                "summary": "Get conversation suggestions",
                "description": "Get suggested conversation starters based on current onboarding step",
                "operationId": "33034da58a73f13f93fccf6b4e66eb49",
                "responses": {
                    "200": {
                        "description": "Suggestions retrieved successfully",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "properties": {
                                        "success": {
                                            "type": "boolean",
                                            "example": true
                                        },
                                        "data": {
                                            "properties": {
                                                "suggestions": {
                                                    "type": "array",
                                                    "items": {
                                                        "type": "string"
                                                    },
                                                    "example": [
                                                        "Hi! I'm Dr. John Doe",
                                                        "I have MBBS and MD degrees"
                                                    ]
                                                },
                                                "current_step": {
                                                    "type": "integer",
                                                    "example": 1
                                                },
                                                "step_title": {
                                                    "type": "string",
                                                    "example": "Basic Information"
                                                }
                                            },
                                            "type": "object"
                                        }
                                    },
                                    "type": "object"
                                }
                            }
                        }
                    },
                    "401": {
                        "description": "Unauthorized",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "$ref": "#/components/schemas/UnauthorizedError"
                                }
                            }
                        }
                    }
                },
                "security": [
                    {
                        "bearerAuth": []
                    }
                ]
            }
        },
...
    },
    "components": {
        "schemas": {
...
            "UnauthorizedError": {
                "properties": {
                    "success": {
                        "type": "boolean",
                        "example": false
                    },
                    "message": {
                        "type": "string",
                        "example": "Unauthorized"
                    }
                },
                "type": "object"
            },
...
        },
        "securitySchemes": {
            "bearerAuth": {
                "type": "http",
                "description": "Enter JWT Bearer token",
                "bearerFormat": "JWT",
                "scheme": "bearer"
            }
        }
    },
...
}